そのメモの作者として認証済みの時のみ削除できるように

+++ laravel/app/Http/Controllers/MemoController.php
@@ -40,9 +40,15 @@ class MemoController extends Controller
     public function delete(Request $request)
     {
         $memo = Memo::find($request['id']);
-        $memo->delete();
-        return response()->json([
-            'message' => 'Deleted successfully :)'
-        ], 200);
+        if (auth()->id() == $memo->user_id) {
+            $memo->delete();
+            return response()->json([
+                'message' => 'Deleted successfully :)'
+            ], 200);
+        } else {
+            return response()->json([
+                'message' => 'Deletion failed :('
+            ], 401);
+        }
     }
 }
